rename byoc to db
