+++ connectors/preflight_check.py
@@ -4,7 +4,7 @@
 # you may not use this file except in compliance with the Elastic License 2.0.
 #
 
-from connectors.byoc import CONNECTORS_INDEX, JOBS_INDEX
+from connectors.db import CONNECTORS_INDEX, JOBS_INDEX
 from connectors.es import ESClient
 from connectors.logger import logger
 from connectors.utils import CancellableSleeps
