+++ connectors/services/job_execution.py
@@ -3,7 +3,7 @@
 # or more contributor license agreements. Licensed under the Elastic License 2.0;
 # you may not use this file except in compliance with the Elastic License 2.0.
 #
-from connectors.byoc import ConnectorIndex, DataSourceError, JobStatus, SyncJobIndex
+from connectors.db import ConnectorIndex, DataSourceError, JobStatus, SyncJobIndex
 from connectors.es.index import DocumentNotFoundError
 from connectors.logger import logger
 from connectors.services.base import BaseService
