+++ connectors/sources/tests/test_mysql.py
@@ -11,7 +11,7 @@ import aiomysql
 import pytest
 from freezegun import freeze_time
 
-from connectors.byoc import Filter
+from connectors.db import Filter
 from connectors.filtering.validation import SyncRuleValidationResult
 from connectors.source import ConfigurableFieldValueError, DataSourceConfiguration
 from connectors.sources.mysql import (
