+++ connectors/tests/test_sync_job_runner.py
@@ -9,7 +9,7 @@ from unittest.mock import ANY, AsyncMock, Mock, patch
 import pytest
 from elasticsearch import ConflictError
 
-from connectors.byoc import Filter, JobStatus, Pipeline
+from connectors.db import Filter, JobStatus, Pipeline
 from connectors.es.index import DocumentNotFoundError
 from connectors.filtering.validation import InvalidFilteringError
 from connectors.sync_job_runner import SyncJobRunner, SyncJobStartError
